retrieval report and file

+++ src/main/java/com/example/accessibility_analyzer/repo/UploadedFileRepo.java
@@ -3,5 +3,5 @@ package com.example.accessibility_analyzer.repo;
 import com.example.accessibility_analyzer.model.UploadedFile;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UploadedFileRepo extends JpaRepository<UploadedFile,Long> {
+public interface UploadedFileRepo extends JpaRepository<UploadedFile,Integer> {
 }
